backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.134.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
motor==3.7.1
pyasn1==0.6.2
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
